_internal/agent/pyproject.toml
[project]
name = "aws-strands-server"
version = "0.1.0"
description = "Strands integration server for AG-UI using OpenAI models"
authors = [{ name = "AG-UI Contributors" }]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "ag-ui-protocol>=0.1.5",
    "fastapi>=0.115.12",
    "uvicorn>=0.34.3",
    "strands-agents[OpenAI]>=1.15.0",
    "strands-agents-tools>=0.2.14",
    "ag_ui_strands~=0.1.0",
    "duckdb>=1.0.0",
    "boto3>=1.35.0",
    "pandas>=2.0.0",
    "requests>=2.31.0",
    "pytz>=2024.1",
    "pyarrow>=15.0.0",
    "browser-cookie3>=0.19.1",
    "requests-negotiate-sspi>=0.5.2; sys_platform == 'win32'",
]
